Fix error reporting in RipROMs and main application.
The RipROMs tool was failing to include a trailing '\n' when reporting
a failure to find !65Host. It's now included.

The main application was ignoring the fact that the Wimp wasn't present
and continued running blindly. We now check for this and report an error
if the Wimp_Initialise fails.